python-services/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP Clients
httpx==0.26.0
requests==2.31.0

# AI Research Services
tavily-python==0.3.0  # AI-powered research
# perplexityai==0.1.0  # Alternative research service (if available)

# Document Processing
pypdf2==3.0.1
python-docx==1.1.0
python-pptx==0.6.23

# LLM Integration
openai==1.10.0
anthropic==0.18.0
langchain==0.1.0
langchain-openai==0.0.5

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Utilities
python-dotenv==1.0.0
tenacity==8.2.3  # Retry logic
validators==0.22.0

# Data Processing
pandas==2.2.0
numpy==1.26.3

# Social Media APIs
tweepy==4.14.0  # Twitter API (optional)

# Image Processing
pillow==10.2.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.3
httpx-mock==0.13.0
